test: add tests for card element, catalog, product cards, product page and fix header
