+++ src/app/components/catalog/__tests__/product_cards.test.ts
@@ -1,0 +1,41 @@
+import Router from '@src/app/router/router';
+
+import ProductCards from '../product-cards/product-cards';
+
+const routes = [
+  { path: '', callBack: jest.fn() },
+  { path: 'login', callBack: jest.fn() },
+  { path: 'registration', callBack: jest.fn() },
+  { path: 'catalog', callBack: jest.fn() },
+  { path: 'product', callBack: jest.fn() },
+  { path: 'error', callBack: jest.fn() },
+];
+
+describe('ProductCards', () => {
+  let products: ProductCards;
+  let router: Router;
+
+  beforeEach(() => {
+    router = new Router(routes);
+    products = new ProductCards(router);
+  });
+
+  it('should show not found title', async () => {
+    const showNotFoundSpy = jest.spyOn(products, 'showNotFoundTitle').mockImplementation(jest.fn());
+
+    const hideBlockPromise = new Promise<void>((resolve) => {
+      products.setProducts([]);
+
+      setTimeout(() => {
+        resolve();
+      }, 200);
+    });
+    await hideBlockPromise;
+
+    products.setProducts([]);
+
+    expect(showNotFoundSpy).toHaveBeenCalled();
+
+    showNotFoundSpy.mockRestore();
+  });
+});
